pyproject.toml
[tool.poetry]
name = "foxbuild"
version = "0.1.0"
description = ""
authors = ["Ivan Filipenkov <hello@vanutp.dev>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pyyaml = "^6.0.1"
pydantic = "^2.7.1"
httpx = "^0.27.0"
pydantic-settings = "^2.2.1"
python-dotenv = "^1.0.1"
joserfc = "^0.9.0"
starlette = "^0.37.2"
uvicorn = "^0.29.0"
yappi = "^1.6.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
